Rewrite Due Reminder search to work on the plist object
This follows an actual understanding of how NSKeyedArchiver
serialisation works. We now use $class to figure out whether
something in the $objects array is a Reminder object.

Having the Reminder objects now, we can dereference UID values
as appropriate, and otherwise directly access values. On top
of that, bplist-parser will serialise large enough numbers into
BigInteger objects, which we now have an awareness of.